+++ psRadiantOne/Public/GlobalNamespace/Get-R1GlobalSpecialAttribute.ps1
@@ -1,0 +1,29 @@
+# .ExternalHelp psRadiantOne-help.xml
+function Get-R1GlobalSpecialAttribute {
+	[CmdletBinding()]
+	[OutputType('psRadiantOne.SpecialAttributesRoot')]
+	param( )
+
+	Begin {
+
+		Assert-R1Session -RequireToken
+
+	}#begin
+
+	Process {
+
+		$URI = Resolve-R1ServiceUrl -Service Namespace -Path 'special_attributes/global'
+
+		$Result = Invoke-R1RestMethod -Uri $URI -Method GET
+
+		if ($null -ne $Result) {
+
+			$Result | Add-CustomType -Type psRadiantOne.SpecialAttributesRoot
+
+		}
+
+	}#process
+
+	End { }#end
+
+}